fixed some stuff and stuff
now you dont go spiderman mode when moving against the cube in the air, i dont know why i never thought to change the friction. added right click to spawn "explosion" thing, having trouble making it so player is repelled when inside of the collider of the explosion. its kind of hard to do and rapier doesnt come with something to do it for me T_T